requirements.txt
flask>=3.0.0
python-dotenv>=1.0.1
openai>=1.35.0
yfinance>=0.2.40
gunicorn>=22.0.0

